# docs/openapi.yaml
openapi: 3.0.3
info:
  title: نظام إدارة محل السبائك والمشغولات — API
  description: |
    REST API لنظام إدارة محل الذهب والفضة: مخزون مركزي، تسعير يومي، فروع متعددة،
    نقاط بيع (POS)، حجوزات، جرد، شيفتات، تقارير، ومزامنة مع WooCommerce.

    ## المصادقة
    - جميع النقاط تحتاج `Authorization: Bearer <token>` ما عدا `/api/auth/login` و `/api/health`.
    - توكن JWT (خوارزمية HS256، صلاحية 12 ساعة) يُحصل عليه من `/api/auth/login`.
    - لكل مسؤول صلاحيات (permissions) تُحمّل مع كائن الموظف عند الدخول.
      النقاط المحمية بصلاحية محددة موثّقة بـ «Requires permission» في الوصف،
      وترجع `403 {error:"forbidden"}` عند غيابها.

    ## الأخطاء
    جميع الأخطاء تُرجع JSON على الصورة: `{"error": "<code>"}` مع اختياري
    `message` (نص عربي قابل للعرض) و `detail` (لأخطاء الاتصال) و `refs` (للحذف المقيّد).
    الكود الداخلي في `error` ثابت وقابل للبرمجة عليه.

    ## توكن الدخول
    استخدم `/api/auth/login` بالحسابات الأولية (PIN: `1234`):
    | username | الدور |
    |---|---|
    | `manager` | مدير المحل |
    | `cashier` | كاشير |
    | `social` | سوشيال |
  version: 1.0.0
  contact:
    name: Jewelry Store System
servers:
  - url: http://localhost:4001
    description: تطوير (Vite proxy)
  - url: https://your-domain.example
    description: إنتاج
...
security:
  - bearerAuth: []
paths:
...
  /api/invoices:
    get:
      tags: [Invoices]
      summary: قائمة الفواتير (آخر 300 مع فلاتر)
      description: |
        `status`: active / returned. `search` يبحث في رقم الفاتورة.
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [active, returned] } }
        - { name: locationId, in: query, schema: { type: integer } }
        - { name: employeeId, in: query, schema: { type: integer } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: search, in: query, schema: { type: string } }
        - { name: method, in: query, schema: { type: string, description: كود طريقة الدفع } }
      responses:
        '200':
          description: الفواتير
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Invoice' }
    post:
      tags: [Invoices]
      summary: إنشاء بيع (POS)
      description: |
        Requires permission: `invoice.create`.
        - يسعّر كل قطعة بسعر اليوم (409 `prices.missing_today` إن لم يوجد سعر).
        - خصم على المصنعية فقط: نسبة أو مبلغ ثابت.
        - تجاوز سقف الخصم للموظف يتطلب `managerPin` (403 `discount.requires_manager`).
        - إن وجد حجز نشط على قطعة مباعة يُكتمل تلقائيًا.
        - رقم الفاتورة تلقائي: `INV-YYYYMMDD-####`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreate'
      responses:
        '201':
          description: الفاتورة الكاملة
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          description: missing:items / bad.quantity / bad.discount
        '403':
          description: discount.disabled_for_cashier / discount.exceeds_cap / discount.requires_manager
        '404':
          description: items.notfound:{id} / customers.notfound
        '409':
          description: items.not_available:{code} / prices.missing_today
  /api/invoices/{id}:
    get:
      tags: [Invoices]
      summary: فاتورة كاملة (مع السطور والمدفوعات)
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: الفاتورة مع `items` و `payments`
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Invoice'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/InvoiceItem' }
                      payments:
                        type: array
                        items: { $ref: '#/components/schemas/Payment' }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/invoices/{id}/return:
    post:
      tags: [Invoices]
      summary: إرجاع / إلغاء فاتورة
      description: |
        Requires permission: `invoice.return`.
        يعيد الكميات إلى المخزون مع حفظ السبب. Idempotent — إرجاع فاتورة
        مُرجعة سابقًا يعيدها كما هي (آمن لتكرار محاولات المزامنة).
      parameters:
        - { $ref: '#/components/parameters/Id' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200':
          description: تم الإرجاع
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }

  # ------------------------------------------------------------------
  # Reservations
  # ------------------------------------------------------------------
...
  /api/customers/{id}/invoices:
    get:
      tags: [Customers]
      summary: فواتير عميل (آخر 100)
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: الفواتير (مع cashierName, locationName)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Invoice' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: توكن JWT من /api/auth/login (صلاحية 12 ساعة)
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
      description: معرّف رقمي
  responses:
...
    NotFound:
      description: غير موجود
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: كود خطأ ثابت قابل للبرمجة، مثال notfound / items.in_use }
        message: { type: string, description: نص عربي قابل للعرض (اختياري) }
        detail: { type: string, description: تفاصيل إضافية (اختياري) }
        refs:
          type: object
          description: إحصاءات المراجع للحذف المقيّد (اختياري)

    # ---- Auth ----
...
    Invoice:
      type: object
      properties:
        id: { type: integer }
        invoiceNo: { type: string, example: INV-20260808-0001 }
        employeeId: { type: integer }
        cashierName: { type: string }
        locationId: { type: integer }
        locationName: { type: string }
        customerId: { type: integer, nullable: true }
        customerName: { type: string, nullable: true }
        customerPhone: { type: string, nullable: true }
        metalSubtotal: { type: number }
        craftsmanshipTotal: { type: number }
        discountAmount: { type: number }
        discountReason: { type: string, nullable: true }
        discountApprovedBy: { type: integer, nullable: true }
        approvedByName: { type: string, nullable: true }
        vatPercent: { type: number }
        vatAmount: { type: number }
        total: { type: number }
        paymentMethod: { type: string }
        paymentMethodName: { type: string }
        paymentMethodColor: { type: string }
        status: { type: string, enum: [active, returned] }
        shiftId: { type: integer, nullable: true }
        isOffline: { type: boolean }
        deviceId: { type: string, nullable: true }
        returnReason: { type: string, nullable: true }
        returnedAt: { type: string, format: date-time, nullable: true }
        returnedBy: { type: integer, nullable: true }
        returnedByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    InvoiceCreate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [itemId]
            properties:
              itemId: { type: integer }
              quantity: { type: integer, minimum: 1, default: 1 }
        locationId: { type: integer, description: افتراضيًا فرع الموظف }
        customerId: { type: integer }
        customerPhone: { type: string }
        paymentMethod: { type: string, default: cash }
        paidAmount: { type: number, description: افتراضيًا الإجمالي }
        discountType: { type: string, enum: [percent, fixed], default: percent }
        discountPercent: { type: number, minimum: 0 }
        discountValue: { type: number, minimum: 0 }
        managerPin: { type: string, description: لتجاوز سقف خصم الكاشير }
        isOffline: { type: boolean, default: false }
        deviceId: { type: string }
    InvoiceItem:
      type: object
      properties:
        id: { type: integer }
        invoiceId: { type: integer }
        itemId: { type: integer }
        quantity: { type: integer }
        itemCodeSnapshot: { type: string }
        itemNameSnapshot: { type: string, nullable: true }
        metalTypeSnapshot: { type: string }
        caratSnapshot: { type: string, nullable: true }
        weightGSnapshot: { type: number }
        metalPriceSnapshot: { type: number }
        metalCostPrice: { type: number, nullable: true }
        craftsmanshipSnapshot: { type: number }
        lineDiscount: { type: number }
        costSnapshot: { type: number, nullable: true }
        lineTotal: { type: number }
    Payment:
      type: object
      properties:
        id: { type: integer }
        invoiceId: { type: integer }
        method: { type: string }
        amount: { type: number }
        receivedBy: { type: integer, nullable: true }
        receivedByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }

    # ---- Reservations ----
...